refactor: avoid circular imports and util fns to util fns
