+++ src/components/app/data/hooks/useCourseMetadata.js
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 import { useParams } from 'react-router-dom';
 
 import useEnterpriseLearner from './useEnterpriseLearner';
-import { queryCourseMetadata } from '../../routes/data/queries';
+import { queryCourseMetadata } from '../queries';
 
 /**
  * Retrieves the course metadata for the given enterprise customer and course key.
