+++ src/components/app/data/hooks/useNotices.js
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useQuery } from '@tanstack/react-query';
 
-import { queryNotices } from '../../routes/data';
+import { queryNotices } from '../queries';
 
 /**
  * Responsible for returning the redirect URL for any notice(s) present
