+++ src/components/app/data/index.js
@@ -1,3 +1,5 @@
 export * from './hooks';
 export * from './utils';
 export * from './constants';
+export * from './services';
+export * from './queries';
