+++ src/components/app/data/queries/subsidies/browseAndRequest.js
@@ -1,5 +1,5 @@
-import { queries } from '../../../../../../utils/queryKeyFactory';
-import { SUBSIDY_REQUEST_STATE } from '../../../../../enterprise-subsidy-requests';
+import queries from '../queryKeyFactory';
+import { SUBSIDY_REQUEST_STATE } from '../../../../enterprise-subsidy-requests';
 
 /**
  * Helper function to assist querying with useQuery package.
