+++ src/components/app/data/queries/subsidies/enterpriseOffers.js
@@ -1,16 +1,18 @@
-import { queries } from '../../../../../utils/queryKeyFactory';
+import queries from '../queryKeyFactory';
 
 /**
  * Helper function to assist querying with useQuery package
  * queries
  * .enterprise
  * .enterpriseCustomer(enterpriseUuid)
- * ._ctx.enrollments
+ * ._ctx.subsidies
+ * ._ctx.enterpriseOffers
  * @returns
  */
-export default function queryEnterpriseCourseEnrollments(enterpriseUuid) {
+export default function queryEnterpriseLearnerOffers(enterpriseUuid) {
   return queries
     .enterprise
     .enterpriseCustomer(enterpriseUuid)
-    ._ctx.enrollments;
+    ._ctx.subsidies
+    ._ctx.enterpriseOffers;
 }
