+++ src/components/app/data/utils.js
@@ -1,7 +1,9 @@
 import dayjs from 'dayjs';
+import { logError } from '@edx/frontend-platform/logging';
 
-import { POLICY_TYPES } from '../../enterprise-user-subsidy/enterprise-offers/data/constants';
+import { ASSIGNMENT_TYPES, POLICY_TYPES } from '../../enterprise-user-subsidy/enterprise-offers/data/constants';
 import { LICENSE_STATUS } from '../../enterprise-user-subsidy/data/constants';
+import { getBrandColorsFromCSSVariables } from '../../../utils/common';
 
 /**
  * Check if system maintenance alert is open, based on configuration.
@@ -79,4 +81,172 @@ export function determineLearnerHasContentAssignmentsOnly({
     && !hasAssignedCodesOrCodeRequests
     && !hasAutoAppliedLearnerCreditPolicies
   );
+}
+
+/**
+ * Helper function to determine which linked enterprise customer user record
+ * should be used for display in the UI.
+ * @param {*} param0
+ * @returns
+ */
+export function determineEnterpriseCustomerUserForDisplay({
+  activeEnterpriseCustomer,
+  activeEnterpriseCustomerUserRoleAssignments,
+  enterpriseSlug,
+  foundEnterpriseCustomerUserForCurrentSlug,
+}) {
+  const activeEnterpriseCustomerUser = {
+    enterpriseCustomer: activeEnterpriseCustomer,
+    roleAssignments: activeEnterpriseCustomerUserRoleAssignments,
+  };
+  if (!enterpriseSlug) {
+    return activeEnterpriseCustomerUser;
+  }
+  if (enterpriseSlug !== activeEnterpriseCustomer.slug && foundEnterpriseCustomerUserForCurrentSlug) {
+    return {
+      enterpriseCustomer: foundEnterpriseCustomerUserForCurrentSlug.enterpriseCustomer,
+      roleAssignments: foundEnterpriseCustomerUserForCurrentSlug.roleAssignments,
+    };
+  }
+  return activeEnterpriseCustomerUser;
+}
+
+/**
+ * Takes a flattened array of assignments and returns an object containing
+ * lists of assignments for each assignment state.
+ *
+ * @param {Array} assignments - List of content assignments.
+ * @returns {{
+*  assignments: Array,
+*  hasAssignments: Boolean,
+*  allocatedAssignments: Array,
+*  hasAllocatedAssignments: Boolean,
+*  canceledAssignments: Array,
+*  hasCanceledAssignments: Boolean,
+*  acceptedAssignments: Array,
+*  hasAcceptedAssignments: Boolean,
+* }}
+*/
+export function getAssignmentsByState(assignments = []) {
+  const allocatedAssignments = [];
+  const acceptedAssignments = [];
+  const canceledAssignments = [];
+  const expiredAssignments = [];
+  const erroredAssignments = [];
+  const assignmentsForDisplay = [];
+
+  assignments.forEach((assignment) => {
+    switch (assignment.state) {
+      case ASSIGNMENT_TYPES.ALLOCATED:
+        allocatedAssignments.push(assignment);
+        break;
+      case ASSIGNMENT_TYPES.ACCEPTED:
+        acceptedAssignments.push(assignment);
+        break;
+      case ASSIGNMENT_TYPES.CANCELED:
+        canceledAssignments.push(assignment);
+        break;
+      case ASSIGNMENT_TYPES.EXPIRED:
+        expiredAssignments.push(assignment);
+        break;
+      case ASSIGNMENT_TYPES.ERRORED:
+        erroredAssignments.push(assignment);
+        break;
+      default:
+        logError(`[getAssignmentsByState] Unsupported state ${assignment.state} for assignment ${assignment.uuid}`);
+        break;
+    }
+  });
+
+  const hasAssignments = assignments.length > 0;
+  const hasAllocatedAssignments = allocatedAssignments.length > 0;
+  const hasAcceptedAssignments = acceptedAssignments.length > 0;
+  const hasCanceledAssignments = canceledAssignments.length > 0;
+  const hasExpiredAssignments = expiredAssignments.length > 0;
+  const hasErroredAssignments = erroredAssignments.length > 0;
+
+  // Concatenate all assignments for display (includes allocated and canceled assignments)
+  assignmentsForDisplay.push(...allocatedAssignments);
+  assignmentsForDisplay.push(...canceledAssignments);
+  assignmentsForDisplay.push(...expiredAssignments);
+  const hasAssignmentsForDisplay = assignmentsForDisplay.length > 0;
+
+  return {
+    assignments,
+    hasAssignments,
+    allocatedAssignments,
+    hasAllocatedAssignments,
+    acceptedAssignments,
+    hasAcceptedAssignments,
+    canceledAssignments,
+    hasCanceledAssignments,
+    expiredAssignments,
+    hasExpiredAssignments,
+    erroredAssignments,
+    hasErroredAssignments,
+    assignmentsForDisplay,
+    hasAssignmentsForDisplay,
+  };
+}
+
+/**
+ * Transform enterprise customer metadata for use by consuming UI components.
+ * @param {Object} enterpriseCustomer
+ * @param {Object} enterpriseFeatures
+ * @returns
+ */
+export function transformEnterpriseCustomer(enterpriseCustomer, enterpriseFeatures) {
+  // If the learner portal is not enabled for the displayed enterprise customer, return null. This
+  // results in the enterprise learner portal not being accessible for the user, showing a 404 page.
+  if (!enterpriseCustomer.enableLearnerPortal) {
+    return null;
+  }
+
+  // Otherwise, learner portal is enabled, so transform the enterprise customer data.
+  const disableSearch = !!(
+    !enterpriseCustomer.enableIntegratedCustomerLearnerPortalSearch
+    && enterpriseCustomer.identityProvider
+  );
+  const showIntegrationWarning = !!(!disableSearch && enterpriseCustomer.identityProvider);
+  const brandColors = getBrandColorsFromCSSVariables();
+  const defaultPrimaryColor = brandColors.primary;
+  const defaultSecondaryColor = brandColors.info100;
+  const defaultTertiaryColor = brandColors.info500;
+  const {
+    primaryColor,
+    secondaryColor,
+    tertiaryColor,
+  } = enterpriseCustomer.brandingConfiguration || {};
+
+  return {
+    ...enterpriseCustomer,
+    brandingConfiguration: {
+      ...enterpriseCustomer.brandingConfiguration,
+      primaryColor: primaryColor || defaultPrimaryColor,
+      secondaryColor: secondaryColor || defaultSecondaryColor,
+      tertiaryColor: tertiaryColor || defaultTertiaryColor,
+    },
+    disableSearch,
+    showIntegrationWarning,
+    enterpriseFeatures,
+  };
+}
+
+/**
+ * Transforms the redeemable policies data by attaching the subsidy expiration date
+ * to each assignment within the policies, if available.
+ * @param {object[]} [policies] - Array of policy objects containing learner assignments.
+ * @returns {object} - Returns modified policies data with subsidy expiration dates attached to assignments.
+ */
+export function transformRedeemablePoliciesData(policies = []) {
+  return policies.map((policy) => {
+    const assignmentsWithSubsidyExpiration = policy.learnerContentAssignments?.map(assignment => ({
+      ...assignment,
+      subsidyExpirationDate: policy.subsidyExpirationDate,
+    }));
+    return {
+      ...policy,
+      learnerContentAssignments: assignmentsWithSubsidyExpiration,
+    };
+  });
 }
