+++ src/components/app/routes/data/index.js
@@ -1,3 +1,1 @@
-export * from './services';
 export * from './utils';
-export * from './queries';
