--- src/components/app/routes/data/queries/canRedeemCourse.js
@@ -1,20 +1,0 @@
-import { getAvailableCourseRuns } from '../../../../course/data/utils';
-import { queries } from '../../../../../utils/queryKeyFactory';
-
-/**
- * Helper function to assist querying with useQuery package
- * queries
- * .enterprise
- * .enterpriseCustomer(enterpriseUuid)
- * ._ctx.course
- * ._ctx.canRedeem(availableCourseRunKeys)
- * @returns
- */
-export default function queryCanRedeem(enterpriseUuid, courseMetadata) {
-  const availableCourseRunKeys = getAvailableCourseRuns(courseMetadata).map(courseRun => courseRun.key);
-  return queries
-    .enterprise
-    .enterpriseCustomer(enterpriseUuid)
-    ._ctx.course
-    ._ctx.canRedeem(availableCourseRunKeys);
-}
