--- src/components/app/routes/data/queries/userEntitlements.js
@@ -1,10 +1,0 @@
-import { queries } from '../../../../../utils/queryKeyFactory';
-
-/**
- * Helper function to assist querying with useQuery package
- * queries.user.entitlements
- * @returns
- */
-export default function queryUserEntitlements() {
-  return queries.user.entitlements;
-}
