+++ src/components/enterprise-user-subsidy/data/constants.js
@@ -2,6 +2,7 @@ export const LICENSE_STATUS = {
   ACTIVATED: 'activated',
   ASSIGNED: 'assigned',
   REVOKED: 'revoked',
+  UNASSIGNED: 'unassigned',
 };
 
 export const LOADING_SCREEN_READER_TEXT = 'loading your edX benefits from your organization';
